backend/requirements.txt
fastapi
uvicorn
motor
python-multipart
google-generativeai
requests
python-dotenv
pandas
